# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "description": "Курсорная пагинация. Заблокированные в выдачу не попадают.",
        "operationId": "UsersController_search",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Подстрока — ищет и в userName, и в fullName («er» найдёт eraj и amERica)",
            "schema": {
              "example": "er",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Поиск пользователей (по userName И fullName, подстрокой)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/by-username/{userName}": {
      "get": {
        "description": "Для @упоминаний и ссылок /u/{userName}: search() ищет подстрокой и по fullName, поэтому точное совпадение им не получить. Заблокированные — 404, как и в поиске.",
        "operationId": "UsersController_findByUserName",
        "parameters": [
          {
            "name": "userName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBriefDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль по точному userName (регистронезависимо)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/suggestions": {
      "get": {
        "description": "Кого читают мои подписки. followedBy → «Подписаны: m.ibrohim».",
        "operationId": "UsersController_suggestions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Рекомендации для вас",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/search-history": {
      "post": {
        "operationId": "UsersController_addSearchText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSearchTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchHistoryItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Добавить текстовый запрос в историю поиска",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_getSearchText",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHistoryItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "История текстовых запросов (с createdAt!)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_clearSearchText",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Очистить историю текстовых запросов",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/search-history/users": {
      "delete": {
        "operationId": "UsersController_clearSearchedUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Очистить историю просмотренных профилей",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_getSearchedUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchedUserItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "История просмотренных профилей (с createdAt!)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/search-history/user/{id}": {
      "delete": {
        "operationId": "UsersController_deleteSearchedUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить один профиль из истории поиска",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/search-history/{id}": {
      "delete": {
        "operationId": "UsersController_deleteSearchText",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить один текстовый запрос из истории",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/search-history/user": {
      "post": {
        "operationId": "UsersController_addSearchedUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSearchedUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchedUserItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Добавить профиль в историю поиска (повтор — поднимает наверх)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me": {
      "delete": {
        "description": "30 дней на восстановление. Все сессии отзываются сразу.",
        "operationId": "UsersController_deleteMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить свой аккаунт (soft-delete)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/report": {
      "post": {
        "operationId": "UsersController_report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreatedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Пожаловаться на пользователя",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/stories/user/{userId}": {
      "get": {
        "operationId": "StoriesController_byUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Истории пользователя",
        "tags": [
          "stories"
        ]
      }
    },
...
    "/api/highlights/user/{userId}": {
      "get": {
        "operationId": "HighlightsController_list",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HighlightDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Актуальное пользователя",
        "tags": [
          "highlights"
        ]
      }
    },
...
    "/api/admin/users/{id}": {
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOkDto"
                }
              }
            }
          },
          "403": {
            "description": "Не ADMIN → 403"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить пользователя (мягко, ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Instagram Backend API",
    "description": "NestJS + Prisma + PostgreSQL. Конверт ответа: { data, errors, statusCode }",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UserBriefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "example": "eraj"
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "userName",
          "fullName",
          "isVerified",
          "isPrivate"
        ]
      },
      "SuggestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "example": "eraj"
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          },
          "followedBy": {
            "example": [
              "m.ibrohim",
              "sadi"
            ],
            "description": "Общие подписки — фронт рисует «Подписаны: m.ibrohim»",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "followedByCount": {
            "type": "number",
            "example": 2,
            "description": "Сколько всего общих подписок"
          }
        },
        "required": [
          "id",
          "userName",
          "fullName",
          "isVerified",
          "isPrivate",
          "followedBy",
          "followedByCount"
        ]
      },
      "AddSearchTextDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "закат",
            "description": "Текст поискового запроса"
          }
        },
        "required": [
          "text"
        ]
      },
      "SearchHistoryItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string",
            "example": "закат"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-15T10:00:00.000Z",
            "description": "Баг softclub #19: createdAt раньше не отдавался — фронт не мог сортировать"
          }
        },
        "required": [
          "id",
          "text",
          "createdAt"
        ]
      },
      "DeletedCountDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "deleted"
        ]
      },
      "AddSearchedUserDto": {
        "type": "object",
        "properties": {
          "searchedUserId": {
            "type": "string",
            "description": "id юзера, чей профиль открыли из поиска"
          }
        },
        "required": [
          "searchedUserId"
        ]
      },
      "SearchedUserItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserBriefDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-15T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "user",
          "createdAt"
        ]
      },
      "AccountDeletedDto": {
        "type": "object",
        "properties": {
          "restorableUntil": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-14T10:00:00.000Z",
            "description": "До этой даты можно восстановить"
          },
          "message": {
            "type": "string",
            "example": "Аккаунт удалён. Восстановить можно в течение 30 дней"
          }
        },
        "required": [
          "restorableUntil",
          "message"
        ]
      },
      "ReportUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Спам и оскорбления"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ReportCreatedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "POST",
              "USER",
              "COMMENT",
              "STORY",
              "CHAT"
            ],
            "example": "USER"
          },
          "targetId": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "example": "Жалоба отправлена, мы её рассмотрим"
          }
        },
        "required": [
          "id",
          "targetType",
          "targetId",
          "message"
        ]
      },
...
      "StoryMusicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 35
          },
          "title": {
            "type": "string",
            "example": "Soundhelix song 1"
          },
          "artist": {
            "type": "string",
            "example": "SoundHelix"
          },
          "streamUrl": {
            "type": "string"
          },
          "coverUrl": {
            "type": "object",
            "nullable": true
          },
          "startSec": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "streamUrl"
        ]
      },
      "StoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "mediaUrl": {
            "type": "string",
            "example": "http://localhost:9000/instagram/images/2026/07/a.webp"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO"
            ],
            "example": "IMAGE"
          },
          "thumbUrl": {
            "type": "object",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "example": 5,
            "description": "Секунд показа"
          },
          "music": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StoryMusicDto"
              }
            ]
          },
          "overlays": {
            "type": "object",
            "nullable": true,
            "description": "overlays (JSON)"
          },
          "filter": {
            "type": "object",
            "nullable": true
          },
          "closeFriendsOnly": {
            "type": "boolean",
            "example": false
          },
          "fromPostId": {
            "type": "object",
            "nullable": true,
            "description": "id поста, если история — репост"
          },
          "isViewed": {
            "type": "boolean",
            "example": false,
            "description": "Смотрел ли Я — считается на СЕРВЕРЕ"
          },
          "isLiked": {
            "type": "boolean",
            "example": false,
            "description": "Лайкнул ли Я"
          },
          "likesCount": {
            "type": "number",
            "example": 3
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mediaUrl",
          "mediaType",
          "duration",
          "closeFriendsOnly",
          "isViewed",
          "isLiked",
          "likesCount",
          "createdAt",
          "expiresAt"
        ]
      },
...
      "HighlightDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Путешествия"
          },
          "coverUrl": {
            "type": "object",
            "nullable": true
          },
          "count": {
            "type": "number",
            "example": 3,
            "description": "Сколько историй в актуальном"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "count",
          "createdAt"
        ]
      },
...
      "AdminOkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  }
}